# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DigitalBalance @home API
  description: |
    API for DigitalBalance @home — a family activity challenge app encouraging
    offline parent-child time through photo-collage challenges.

    **Spec-driven workflow:** This YAML is the authoritative contract.
    Pydantic schemas in `server/app/schemas/generated.py` are generated from it
    via `datamodel-codegen`. Never edit `generated.py` by hand.
  version: 0.1.0
  contact:
    name: HIP26 — DigitalBalance @home
    email: ignacio.garcian15@gmail.com
servers:
  - url: http://localhost:8000
    description: Local Docker Compose
security:
  - BearerAuth: []
...
paths:
...
  /photos:
    post:
      operationId: uploadPhoto
      tags: [Photos]
      summary: Upload a photo for an activity completion (returns 202; compression is async)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PhotoUploadRequest"
      responses:
        "202":
          description: Upload accepted; compression is processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /photos/{completionId}/url:
    get:
      operationId: getPhotoUrl
      tags: [Photos]
      summary: Get a pre-signed S3 URL for a completion photo (15-minute TTL)
      parameters:
        - $ref: "#/components/parameters/CompletionId"
      responses:
        "200":
          description: Pre-signed URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUrlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Completions ─────────────────────────────────────────────
  /completions:
    post:
      operationId: createCompletion
      tags: [Completions]
      summary: Mark an activity as self-reported complete (no photo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompletionRequest"
      responses:
        "201":
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Completion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /completions/me:
    get:
      operationId: getMyCompletions
      tags: [Completions]
      summary: Get the family's completion history across all challenges
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Completions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompletionHistoryItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /completions/{completionId}:
    get:
      operationId: getCompletion
      tags: [Completions]
      summary: Get a single completion (used for polling photo compression status)
      parameters:
        - $ref: "#/components/parameters/CompletionId"
      responses:
        "200":
          description: Completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Completion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCompletion
      tags: [Completions]
      summary: Delete a completion and its associated S3 photo (family-scoped)
      parameters:
        - $ref: "#/components/parameters/CompletionId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /completions/{completionId}/photo:
    patch:
      operationId: reuploadCompletionPhoto
      tags: [Completions]
      summary: Replace the photo on an existing completion
      description: >
        On a verified completion, updates only the stored photo (status and
        points unchanged). On a rejected or pending completion, re-triggers
        the compression pipeline and resets status to processing (then
        pending_verification). Not permitted on self_reported completions.
      parameters:
        - $ref: "#/components/parameters/CompletionId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PhotoReuploadRequest"
      responses:
        "202":
          description: Photo accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReuploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Verification ────────────────────────────────────────────
...
  /time-spent:
    get:
      operationId: getTimeSpentInsight
      tags: [Time Spent]
      summary: Get the authenticated parent's weekly or monthly time-spent insight
      description: >
        Combines activity durations attributed to the authenticated parent with
        that parent's manual daily time. No user or family identifier is accepted.
      parameters:
        - name: period
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/TimeSpentPeriod"
        - name: anchor_date
          in: query
          required: true
          description: Parent's local calendar date anchoring the requested week or month.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Personal time-spent series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSpentInsight"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /time-spent/manual:
    put:
      operationId: upsertManualTime
      tags: [Time Spent]
      summary: Set or replace the authenticated parent's manual time for a local date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertManualTimeRequest"
      responses:
        "200":
          description: Manual value created or replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualTimeEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Journal (retained, frontend phased out) ────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ──────────────────────────────────────────────
  parameters:
...
    CompletionId:
      name: completionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        Preferred content language for curated activity and challenge text.
        Recognises "de" and "en" (default "de"). When "en" is requested, curated
        content is returned in English where a translation exists; otherwise the
        original/German text is returned. Does not affect user-created content.
      schema:
        type: string
        example: en

  # ── Common Responses ────────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorised for this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ── Schemas ─────────────────────────────────────────────────
  schemas:
...
    CompletionStatus:
      type: string
      enum: [processing, pending_verification, verified, rejected, self_reported]
...
    TimeSpentPeriod:
      type: string
      enum: [weekly, monthly]

    # Shared
    ErrorResponse:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
        code:
          type: string
...
    PhotoUploadRequest:
      type: object
      required: [challenge_activity_id, image, completed_on]
      properties:
        challenge_activity_id:
          type: string
          format: uuid
        image:
          type: string
          format: binary
          description: JPEG or PNG; maximum 10 MB
        caption:
          type: string
          nullable: true
        shared_to_feed:
          type: boolean
          default: false
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 1440
          nullable: true
          description: >
            Minutes the family reports spending on the activity, from a fixed
            dropdown. Required when the activity's resolved point tier is
            casual (server-enforced); optional otherwise.
        completed_on:
          type: string
          format: date
          description: Parent's local calendar date when the activity was completed
    PhotoUploadResponse:
      type: object
      required: [completion_id]
      properties:
        completion_id:
          type: string
          format: uuid
    PhotoUrlResponse:
      type: object
      required: [url, expires_at]
      properties:
        url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time

    # Completions
    Completion:
      type: object
      required: [id, challenge_activity_id, family_id, completed_by_user_id, status, completed_on, shared_to_feed, completed_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        challenge_activity_id:
          type: string
          format: uuid
        family_id:
          type: string
          format: uuid
        completed_by_user_id:
          type: string
          format: uuid
          description: The parent who uploaded the photo or marked it complete
        status:
          $ref: "#/components/schemas/CompletionStatus"
        photo_url:
          type: string
          format: uri
          nullable: true
          description: Pre-signed S3 URL; present when status is pending_verification, verified, or rejected
        caption:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
          description: Set when status is rejected; null otherwise
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 1440
          nullable: true
          description: Minutes the family reported spending on the activity
        completed_on:
          type: string
          format: date
          description: Parent's local calendar date when the activity was completed
        shared_to_feed:
          type: boolean
        completed_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCompletionRequest:
      type: object
      required: [challenge_activity_id, duration_minutes, completed_on]
      properties:
        challenge_activity_id:
          type: string
          format: uuid
        caption:
          type: string
          nullable: true
        shared_to_feed:
          type: boolean
          default: false
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 1440
          description: Required time spent for a self-reported completion
        completed_on:
          type: string
          format: date
          description: Parent's local calendar date when the activity was completed
    CompletionHistoryItem:
      type: object
      required: [id, activity_id, activity_title, challenge_title, status, completed_on, completed_at]
      properties:
        id:
          type: string
          format: uuid
        activity_id:
          type: string
          format: uuid
          description: The underlying activity, stable across challenges and languages
        activity_title:
          type: string
        challenge_title:
          type: string
        status:
          $ref: "#/components/schemas/CompletionStatus"
        photo_url:
          type: string
          format: uri
          nullable: true
        caption:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
          description: Set when status is rejected; null otherwise
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 1440
          nullable: true
        completed_on:
          type: string
          format: date
        completed_at:
          type: string
          format: date-time
    PhotoReuploadRequest:
      type: object
      required: [image]
      properties:
        image:
          type: string
          format: binary
          description: JPEG or PNG; maximum 10 MB
    ReuploadResponse:
      type: object
      required: [completion_id, status]
      properties:
        completion_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CompletionStatus"

    # Verification
...
    UpsertManualTimeRequest:
      type: object
      required: [entry_date, minutes]
      additionalProperties: false
      properties:
        entry_date:
          type: string
          format: date
          description: Parent's local calendar date; active UI sends local today
        minutes:
          type: integer
          minimum: 1
          maximum: 1440
    ManualTimeEntry:
      type: object
      required: [id, entry_date, minutes, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        entry_date:
          type: string
          format: date
        minutes:
          type: integer
          minimum: 1
          maximum: 1440
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DailyTimeTotal:
      type: object
      required: [date, activity_minutes, manual_minutes, total_minutes]
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
        activity_minutes:
          type: integer
          minimum: 0
        manual_minutes:
          type: integer
          minimum: 0
        total_minutes:
          type: integer
          minimum: 0
    WeeklyTimeTotal:
      type: object
      required: [start_date, end_date, total_minutes]
      additionalProperties: false
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        total_minutes:
          type: integer
          minimum: 0
    TimeSpentInsight:
      type: object
      required: [period, range_start, range_end, elapsed_end, daily_totals, weekly_totals, average_weekly_minutes]
      additionalProperties: false
      properties:
        period:
          $ref: "#/components/schemas/TimeSpentPeriod"
        range_start:
          type: string
          format: date
          description: Start of the full displayed week/month, regardless of today's date
        range_end:
          type: string
          format: date
          description: End of the full displayed week/month, regardless of today's date — may be in the future
        elapsed_end:
          type: string
          format: date
          description: min(range_end, today) — the last non-future date in range; use to exclude not-yet-happened days from averages
        daily_totals:
          type: array
          description: One entry per day in the full range; days after elapsed_end are always zero
          items:
            $ref: "#/components/schemas/DailyTimeTotal"
        weekly_totals:
          type: array
          description: Empty for weekly view; one Monday-based bucket per calendar week in the month for monthly view, clipped only at month boundaries (not at today)
          items:
            $ref: "#/components/schemas/WeeklyTimeTotal"
        average_weekly_minutes:
          anyOf:
            - type: integer
              minimum: 0
            - type: "null"
          description: Null for weekly view; rounded average of monthly weekly totals

    # Journal
...